Working on implementing plus
Hopefully it works this time. This time I'm trying to make sure to take into account the fact that the last operator is most likely not the same as the current operator. Thus, chainign of operations should work with this method, whereas it did not with the old one. While this may not be the most efficient method to solve the problem, it should work, unlike the last one.