Implemented all operators.
All operators now work properly both on their own and when chained together. They will work properly up until the equal sign in pressed. All that is left to implement functionality wise is the backspace and the decimal point. After that I will go on to work on trying to fully implement the screen so that wrapping will take place, as well as make the calculator itself look at least halfway decent and not like a christmas demon